Time: 6 ms (89.05%) | Memory: 82.6 MB (94.63%) - LeetSync

+++ 1961-maximum-ice-cream-bars/maximum-ice-cream-bars.java
@@ -15,7 +15,7 @@ class Solution {
         long cur = coins;
         int i = 0, res = 0;
         while (i <= mx && cur > 0) {
-            long req = i * cnt[i];
+            long req = (long) i * cnt[i];
 
             if (cur - req < 0) {
                 res += (int) (cur / i);
